requirements.txt
django
djangorestframework
pillow
django-cors-headers
psycopg2
python-decouple